vibe
create a NW.js app, using MUI for the UI, react-router for routing and node-json-db to store data, the app has a MUI permanent navigation drawer on the left, in the navigation panel there is an "manage assets" button that navigates to the manage assets page. On the manage assets page, create  a CRUD that lets the user add/delete/list one of the catagory BTC, ETH, ADA and SUI. Under each category let the user add/delete/list address strings.