vibe
Add a new page "transactions". On this page, there is header "Polygon USDT". Under this header, there is a text input box for address with the default value "0x378b6f2610526217122eced61f9d64097eb58010" and a button "query transactions". When the button is clicked, run getErc20Transactions in fetch.js with chainId '137',  tokenAddress '0xc2132d05d31c914a87c6611c10748aeb04b58e8f' and the input address. Output the results as a table. The column of the table are:
-hash
- date: in dd/mm/yyyy format, converted from the timestamp value, which is seconds after epoch
- value: adjusted by tokenDecimal.
If the value of from is the input address, the value is in a parenthesis in red. If the value of to is the input address, the value is in black. Otherwise show a '?'.