vibe
Add a date picker to allow select one month, default to last month. Add a query button, when clicked, get first and last epoch of the selected month using getEpochsForMonth function in cardanoRewards.ts. Then for each selected stake address, call getRewardHistory function in cardanoRewards. Filter the results by the epochs of the picked month and show the rewards by epoch. The amount in the result returned by getRewardHistory is in lovelace and should be converted to ADA. Also convert the ADA value to USD using fetch.ts.